azer

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -10,7 +10,7 @@
 
 void print_number(int n)
 {
-	long int num;
+	int num;
 
 	if (n < 0)
 	{
